Fix review comments: syntax, error handling, contracts, adapter correctness
- Fix unescaped apostrophes in who-fucked-up causing svelte-check failure
- Wrap all async handlers in try/catch/finally so loading state always clears
- Surface packaging errors and null-filter image previews in all 3 mode pages
- Centralize :global(body) base styles in layout; remove per-page body overrides
- Add .int() constraint to rating fields in meechie-tool and meechie-voice contracts
- Add ExcuseRatingFallbackSchema (keywords.min(0)) to support empty-keyword fallback
- Wire rateExcuse fallback to pack.responses.excuseRatingFallback instead of inline literal
- Fix random_meechie headline to static label 'Random Meechie'; response holds the saying
- Remove overly broad bare 'working' keyword from working-late excuse rating entry
- Update meechie-voice fixture to match adapter state exactly (all fields, correct keywords)
- Update nav links and stale UI copy to match current tool names

+++ contracts/meechie-tool.contract.ts
@@ -108,7 +108,7 @@ export const MeechieToolOutputSchema = z.object({
 	toolId: MeechieToolIdSchema,
 	headline: NonEmptyStringSchema,
 	response: NonEmptyStringSchema,
-	rating: z.number().min(1).max(10).optional()
+	rating: z.number().int().min(1).max(10).optional()
 });
 
 export const MeechieToolResultSchema = resultSchema(MeechieToolOutputSchema);
